requirements.txt
torch
pillow
torchvision
streamlit